+++ aimmd/ops/__init__.py
@@ -17,4 +17,4 @@ along with AIMMD. If not, see <https://www.gnu.org/licenses/>.
 from .selector import RCModelSelector
 #from .traininghook import TrainingHook as TrainingHookLegacy
 from .hooks import TrainingHook, AimmdStorageHook, DensityCollectionHook
-from .utils import set_rcmodel_in_all_selectors
+from .utils import set_rcmodel_and_hook_in_all_selectors